feat(multi_object_tracker): debug timer to work with reduced replay rate recomputation (#6706)
* fix: mot debug timer, full set of timings


* fix: reduce timers, rename topics


* fix: separate debugger


* chore: fix comments


* fix: fix stamp initialization, remove unused function


---------
