+++ perception/multi_object_tracker/src/multi_object_tracker_core.cpp
@@ -66,113 +66,6 @@ boost::optional<geometry_msgs::msg::Transform> getTransformAnonymous(
 
 }  // namespace
 
-TrackerDebugger::TrackerDebugger(rclcpp::Node & node)
-: diagnostic_updater_(&node), node_(node), last_input_stamp_(node.now())
-{
-  // declare debug parameters to decide whether to publish debug topics
-  loadParameters();
-  // initialize debug publishers
-  stop_watch_ptr_ = std::make_unique<tier4_autoware_utils::StopWatch<std::chrono::milliseconds>>();
-  if (debug_settings_.publish_processing_time) {
-    processing_time_publisher_ =
-      std::make_unique<tier4_autoware_utils::DebugPublisher>(&node_, "multi_object_tracker");
-  }
-
-  if (debug_settings_.publish_tentative_objects) {
-    debug_tentative_objects_pub_ =
-      node_.create_publisher<autoware_auto_perception_msgs::msg::TrackedObjects>(
-        "debug/tentative_objects", rclcpp::QoS{1});
-  }
-
-  // initialize stop watch and diagnostics
-  startStopWatch();
-  setupDiagnostics();
-}
-
-void TrackerDebugger::setupDiagnostics()
-{
-  diagnostic_updater_.setHardwareID(node_.get_name());
-  diagnostic_updater_.add(
-    "Perception delay check from original header stamp", this, &TrackerDebugger::checkDelay);
-  diagnostic_updater_.setPeriod(0.1);
-}
-
-void TrackerDebugger::loadParameters()
-{
-  try {
-    debug_settings_.publish_processing_time =
-      node_.declare_parameter<bool>("publish_processing_time");
-    debug_settings_.publish_tentative_objects =
-      node_.declare_parameter<bool>("publish_tentative_objects");
-    debug_settings_.diagnostics_warn_delay =
-      node_.declare_parameter<double>("diagnostics_warn_delay");
-    debug_settings_.diagnostics_error_delay =
-      node_.declare_parameter<double>("diagnostics_error_delay");
-  } catch (const std::exception & e) {
-    RCLCPP_WARN(node_.get_logger(), "Failed to declare parameter: %s", e.what());
-    debug_settings_.publish_processing_time = false;
-    debug_settings_.publish_tentative_objects = false;
-    debug_settings_.diagnostics_warn_delay = 0.5;
-    debug_settings_.diagnostics_error_delay = 1.0;
-  }
-}
-
-void TrackerDebugger::checkDelay(diagnostic_updater::DiagnosticStatusWrapper & stat)
-{
-  const double delay = elapsed_time_from_sensor_input_;  // [s]
-
-  if (delay == 0.0) {
-    stat.summary(diagnostic_msgs::msg::DiagnosticStatus::OK, "Detection delay is not calculated.");
-  } else if (delay < debug_settings_.diagnostics_warn_delay) {
-    stat.summary(diagnostic_msgs::msg::DiagnosticStatus::OK, "Detection delay is acceptable");
-  } else if (delay < debug_settings_.diagnostics_error_delay) {
-    stat.summary(
-      diagnostic_msgs::msg::DiagnosticStatus::WARN, "Detection delay is over warn threshold.");
-  } else {
-    stat.summary(
-      diagnostic_msgs::msg::DiagnosticStatus::ERROR, "Detection delay is over error threshold.");
-  }
-
-  stat.add("Detection delay", delay);
-}
-
-void TrackerDebugger::publishProcessingTime()
-{
-  const double processing_time_ms = stop_watch_ptr_->toc("processing_time", true);
-  const double cyclic_time_ms = stop_watch_ptr_->toc("cyclic_time", true);
-  const auto current_time = node_.now();
-  elapsed_time_from_sensor_input_ = (current_time - last_input_stamp_).seconds();
-  if (debug_settings_.publish_processing_time) {
-    processing_time_publisher_->publish<tier4_debug_msgs::msg::Float64Stamped>(
-      "debug/cyclic_time_ms", cyclic_time_ms);
-    processing_time_publisher_->publish<tier4_debug_msgs::msg::Float64Stamped>(
-      "debug/processing_time_ms", processing_time_ms);
-    processing_time_publisher_->publish<tier4_debug_msgs::msg::Float64Stamped>(
-      "debug/pipeline_latency_ms", elapsed_time_from_sensor_input_ * 1e3);
-  }
-}
-
-void TrackerDebugger::publishTentativeObjects(
-  const autoware_auto_perception_msgs::msg::TrackedObjects & tentative_objects) const
-{
-  if (debug_settings_.publish_tentative_objects) {
-    debug_tentative_objects_pub_->publish(tentative_objects);
-  }
-}
-
-void TrackerDebugger::startStopWatch()
-{
-  stop_watch_ptr_->tic("cyclic_time");
-  stop_watch_ptr_->tic("processing_time");
-}
-
-void TrackerDebugger::startMeasurementTime(const rclcpp::Time & measurement_header_stamp)
-{
-  last_input_stamp_ = measurement_header_stamp;
-  // start measuring processing time
-  stop_watch_ptr_->toc("processing_time", true);
-}
-
 MultiObjectTracker::MultiObjectTracker(const rclcpp::NodeOptions & node_options)
 : rclcpp::Node("multi_object_tracker", node_options),
   tf_buffer_(this->get_clock()),
@@ -243,7 +136,7 @@ void MultiObjectTracker::onMeasurement(
   const autoware_auto_perception_msgs::msg::DetectedObjects::ConstSharedPtr input_objects_msg)
 {
   /* keep the latest input stamp and check transform*/
-  debugger_->startMeasurementTime(rclcpp::Time(input_objects_msg->header.stamp));
+  debugger_->startMeasurementTime(this->now(), rclcpp::Time(input_objects_msg->header.stamp));
   const auto self_transform = getTransformAnonymous(
     tf_buffer_, "base_link", world_frame_id_, input_objects_msg->header.stamp);
   if (!self_transform) {
@@ -472,7 +365,7 @@ void MultiObjectTracker::publish(const rclcpp::Time & time) const
   published_time_publisher_->publish_if_subscribed(tracked_objects_pub_, output_msg.header.stamp);
 
   // Debugger Publish if enabled
-  debugger_->publishProcessingTime();
+  debugger_->endPublishTime(this->now(), time);
   debugger_->publishTentativeObjects(tentative_objects_msg);
 }
